+++ src/main/java/db/demo/dao/RestDAO.java
@@ -69,7 +69,7 @@ public class RestDAO {
         	connection = JdbcUtils.getconn();
         	String sql = "INSERT INTO restaurant_info(Rest_Id, Rest_Name, Rest_Address, Description) VALUES(?, ?, ?, ?);";
         	preparedStatement = connection.prepareStatement(sql);
-        	preparedStatement.setInt(1, rest.getRestID());
+        	preparedStatement.setLong(1, rest.getRestID());
         	preparedStatement.setString(2, rest.getRestName());
         	preparedStatement.setString(3, rest.getRestAddress());
         	preparedStatement.setString(4, rest.getDescription());
@@ -87,7 +87,7 @@ public class RestDAO {
 
     // 利用 restID 刪除 合作餐廳
     // 已測試成功
-    public static void delRest(int restID)
+    public static void delRest(Long restID)
     {
     	Connection connection = null;
         PreparedStatement preparedStatement = null;
@@ -95,7 +95,7 @@ public class RestDAO {
         	connection = JdbcUtils.getconn();
         	String sql = "DELETE FROM restaurant_info WHERE Rest_Id = ?";
         	preparedStatement = connection.prepareStatement(sql);
-        	preparedStatement.setInt(1, restID);
+        	preparedStatement.setLong(1, restID);
         	preparedStatement.executeUpdate();
         }
         catch(SQLException e) 
@@ -118,9 +118,9 @@ public class RestDAO {
         	connection = JdbcUtils.getconn();
         	String sql = "INSERT INTO meal(Food_Id, Food_Name, Rest_Id, Cost, Description, Image) VALUES(?, ?, ?, ?, ?, ?);";
         	preparedStatement = connection.prepareStatement(sql);
-        	preparedStatement.setInt(1, menu.getFood_Id());
+        	preparedStatement.setLong(1, menu.getFood_Id());
         	preparedStatement.setString(2, menu.getFood_Name());
-        	preparedStatement.setInt(3, menu.getRest_Id());
+        	preparedStatement.setLong(3, menu.getRest_Id());
         	preparedStatement.setInt(4, menu.getCost());
         	preparedStatement.setString(5, menu.getDescription());
         	preparedStatement.setString(6, menu.getImage());
@@ -138,7 +138,7 @@ public class RestDAO {
 
     // 利用 restID 和 foodID 刪除 指定合作餐廳 的菜單
     // 已測試成功
-    public static void delRestMenu(int foodID)
+    public static void delRestMenu(Long foodID)
     {
     	Connection connection = null;
         PreparedStatement preparedStatement = null;
@@ -146,7 +146,7 @@ public class RestDAO {
         	connection = JdbcUtils.getconn();
         	String sql = "DELETE FROM meal WHERE Food_Id = ?";
         	preparedStatement = connection.prepareStatement(sql);
-        	preparedStatement.setInt(1, foodID);
+        	preparedStatement.setLong(1, foodID);
         	preparedStatement.executeUpdate();
         }
         catch(SQLException e)
