+++ src/main/java/order/controller/service/ShowCommonUserCurrentOrderServlet.java
@@ -26,9 +26,9 @@ public class ShowCommonUserCurrentOrderServlet extends HttpServlet {
         Gson gson = new GsonBuilder().disableHtmlEscaping().setFieldNamingPolicy(FieldNamingPolicy.IDENTITY).create();
         String parm = request.getParameter("userID");
         String json = null;
-        int userID;
+        Long userID;
         try {
-            userID = (int)Long.parseLong(parm);
+            userID = Long.parseLong(parm);
             if(UserDAO.showUserIdentity(userID).equals(MemberSetting.UserStatus.CUSTOMER)) {
                 // 查詢 食客 當前訂單
                 json = gson.toJson(OrderDAO.searchEaterOrder(userID));
